+++ backend/services/solid/notifications-provider.js
@@ -1,8 +1,8 @@
-const { NotificationProviderService } = require('@semapps/solid');
+const { NotificationsProviderService } = require('@semapps/solid');
 const CONFIG = require('../../config/config');
 
 module.exports = {
-  mixins: [NotificationProviderService],
+  mixins: [NotificationsProviderService],
   settings: {
     baseUrl: CONFIG.BASE_URL,
     queueServiceUrl: CONFIG.QUEUE_SERVICE_URL
