+++ config/docker/arena-test.json
@@ -5,35 +5,35 @@
       "hostId": "test-pod-provider",
       "host": "redis",
       "port": 6379,
-      "db": 0
+      "db": 11
     },
     {
       "name": "remotePost",
       "hostId": "test-pod-provider",
       "host": "redis",
       "port": 6379,
-      "db": 0
+      "db": 11
     },
     {
       "name": "processInbox",
       "hostId": "test-pod-provider",
       "host": "redis",
       "port": 6379,
-      "db": 0
+      "db": 11
     },
     {
       "name": "processOutbox",
       "hostId": "test-pod-provider",
       "host": "redis",
       "port": 6379,
-      "db": 0
+      "db": 11
     },
     {
       "name": "webhookPost",
       "hostId": "test-pod-provider",
       "host": "redis",
       "port": 6379,
-      "db": 0
+      "db": 11
     },
     {
       "name": "remotePost",
