+++ packages/app/package.json
@@ -11,7 +11,6 @@
     "directory": "packages/app"
   },
   "dependencies": {
-    "@activitypods/core": "2.0.0-alpha.11",
     "@rdfjs/data-model": "^1.3.4",
     "@semapps/activitypub": "0.9.1",
     "@semapps/ldp": "0.9.1",
