+++ tests/app-installation.test.js
@@ -2,7 +2,7 @@ const urlJoin = require('url-join');
 const waitForExpect = require('wait-for-expect');
 const { MIME_TYPES } = require('@semapps/mime-types');
 const { arrayOf } = require('@semapps/ldp');
-const { initializePodProvider, initializeAppServer, installApp } = require('./initialize');
+const { connectPodProvider, clearAllData, initializeAppServer, installApp } = require('./initialize');
 const ExampleAppService = require('./apps/example.app');
 const { ACTIVITY_TYPES } = require('@semapps/activitypub');
 
@@ -26,7 +26,9 @@ describe('Test app installation', () => {
     appRegistration;
 
   beforeAll(async () => {
-    podProvider = await initializePodProvider();
+    await clearAllData();
+
+    podProvider = await connectPodProvider();
 
     const actorData = require(`./data/actor1.json`);
     const { webId } = await podProvider.call('auth.signup', actorData);
@@ -131,19 +133,21 @@ describe('Test app installation', () => {
   });
 
   test('User installs app and grants all access needs', async () => {
-    await alice.call('activitypub.outbox.post', {
-      collectionUri: alice.outbox,
-      type: 'apods:Install',
-      object: APP_URI,
-      'apods:acceptedAccessNeeds': [
-        requiredAccessNeedGroup['interop:hasAccessNeed'],
-        optionalAccessNeedGroup['interop:hasAccessNeed']
-      ],
-      'apods:acceptedSpecialRights': [
-        requiredAccessNeedGroup['apods:hasSpecialRights'],
-        optionalAccessNeedGroup['apods:hasSpecialRights']
-      ]
-    });
+    await expect(
+      alice.call('activitypub.outbox.post', {
+        collectionUri: alice.outbox,
+        type: 'apods:Install',
+        object: APP_URI,
+        'apods:acceptedAccessNeeds': [
+          requiredAccessNeedGroup['interop:hasAccessNeed'],
+          optionalAccessNeedGroup['interop:hasAccessNeed']
+        ],
+        'apods:acceptedSpecialRights': [
+          requiredAccessNeedGroup['apods:hasSpecialRights'],
+          optionalAccessNeedGroup['apods:hasSpecialRights']
+        ]
+      })
+    ).resolves.not.toThrow();
 
     let appRegistrationUri, creationActivityUri;
 
@@ -492,14 +496,16 @@ describe('Test app installation', () => {
   });
 
   test('User uninstalls app', async () => {
-    await alice.call('activitypub.outbox.post', {
-      collectionUri: alice.outbox,
-      type: ACTIVITY_TYPES.UNDO,
-      object: {
-        type: 'apods:Install',
-        object: APP_URI
-      }
-    });
+    await expect(
+      alice.call('activitypub.outbox.post', {
+        collectionUri: alice.outbox,
+        type: ACTIVITY_TYPES.UNDO,
+        object: {
+          type: 'apods:Install',
+          object: APP_URI
+        }
+      })
+    ).resolves.not.toThrow();
 
     let appRegistrationUri;
 
@@ -604,31 +610,30 @@ describe('Test app installation', () => {
   });
 
   test('User installs app and do not grant required access needs', async () => {
-    const [creationActivityUri, appRegistrationUri] = await installApp(
+    const createRegistrationActivity = await installApp(
       alice,
       APP_URI,
+      false,
       optionalAccessNeedGroup['interop:hasAccessNeed'],
       optionalAccessNeedGroup['apods:hasSpecialRights']
     );
 
-    await waitForExpect(async () => {
-      const inbox = await alice.call('activitypub.collection.get', {
-        resourceUri: alice.inbox,
-        page: 1
-      });
-
-      expect(inbox?.orderedItems[0]).toMatchObject({
-        type: ACTIVITY_TYPES.REJECT,
-        object: creationActivityUri,
-        summary: 'One or more required access needs have not been granted'
-      });
-    });
+    await expect(
+      alice.call('activitypub.inbox.awaitActivity', {
+        collectionUri: alice.inbox,
+        matcher: {
+          type: ACTIVITY_TYPES.REJECT,
+          object: createRegistrationActivity.id,
+          summary: 'One or more required access needs have not been granted'
+        }
+      })
+    ).resolves.not.toThrow();
 
     // The ApplicationRegistration should be deleted
     await waitForExpect(async () => {
       await expect(
         alice.call('ldp.resource.get', {
-          resourceUri: appRegistrationUri,
+          resourceUri: createRegistrationActivity.object,
           accept: MIME_TYPES.JSON
         })
       ).rejects.toThrow();
@@ -636,24 +641,15 @@ describe('Test app installation', () => {
   });
 
   test('User installs app and only grant required access needs', async () => {
-    const [creationActivityUri] = await installApp(
-      alice,
-      APP_URI,
-      requiredAccessNeedGroup['interop:hasAccessNeed'],
-      requiredAccessNeedGroup['apods:hasSpecialRights']
-    );
-
-    await waitForExpect(async () => {
-      const inbox = await alice.call('activitypub.collection.get', {
-        resourceUri: alice.inbox,
-        page: 1
-      });
-
-      expect(inbox?.orderedItems[0]).toMatchObject({
-        type: ACTIVITY_TYPES.ACCEPT,
-        object: creationActivityUri
-      });
-    });
+    await expect(
+      installApp(
+        alice,
+        APP_URI,
+        true,
+        requiredAccessNeedGroup['interop:hasAccessNeed'],
+        requiredAccessNeedGroup['apods:hasSpecialRights']
+      )
+    ).resolves.not.toThrow();
   });
 
   test('User installs app 2 which has same requirements', async () => {
