+++ tests/pod-resources.test.js
@@ -1,8 +1,7 @@
-const path = require('path');
 const urlJoin = require('url-join');
 const { MIME_TYPES } = require('@semapps/mime-types');
 const { triple, namedNode, literal } = require('@rdfjs/data-model');
-const { initialize, initializeAppServer, clearDataset, listDatasets, installApp } = require('./initialize');
+const { connectPodProvider, clearAllData, initializeAppServer, installApp } = require('./initialize');
 const ExampleAppService = require('./apps/example.app');
 
 jest.setTimeout(80000);
@@ -14,7 +13,7 @@ const APP_URI = urlJoin(APP_SERVER_BASE_URL, 'app');
 
 describe('Test Pod resources handling', () => {
   let actors = [],
-    podServer,
+    podProvider,
     alice,
     bob,
     appServer,
@@ -23,22 +22,17 @@ describe('Test Pod resources handling', () => {
     bobNoteUri;
 
   beforeAll(async () => {
-    const datasets = await listDatasets();
-    for (let dataset of datasets) {
-      await clearDataset(dataset);
-    }
+    await clearAllData();
 
-    podServer = await initialize(3000, 'settings');
-    podServer.loadService(path.resolve(__dirname, './services/profiles.app.js'));
-    await podServer.start();
+    podProvider = await connectPodProvider();
 
     appServer = await initializeAppServer(3001, 'appData', 'app_settings', 1, ExampleAppService);
     await appServer.start();
 
     for (let i = 1; i <= NUM_PODS; i++) {
       const actorData = require(`./data/actor${i}.json`);
-      const { webId } = await podServer.call('auth.signup', actorData);
-      actors[i] = await podServer.call(
+      const { webId } = await podProvider.call('auth.signup', actorData);
+      actors[i] = await podProvider.call(
         'activitypub.actor.awaitCreateComplete',
         {
           actorUri: webId,
@@ -47,7 +41,7 @@ describe('Test Pod resources handling', () => {
         { meta: { dataset: actorData.username } }
       );
       actors[i].call = (actionName, params, options = {}) =>
-        podServer.call(actionName, params, {
+        podProvider.call(actionName, params, {
           ...options,
           meta: { ...options.meta, webId, dataset: actors[i].preferredUsername }
         });
@@ -62,7 +56,7 @@ describe('Test Pod resources handling', () => {
 
   afterAll(async () => {
     await appServer.stop();
-    await podServer.stop();
+    await podProvider.stop();
   });
 
   test('Get local data through app', async () => {
