+++ tests/pods-creation.test.js
@@ -1,7 +1,7 @@
 const urlJoin = require('url-join');
 const waitForExpect = require('wait-for-expect');
 const { MIME_TYPES } = require('@semapps/mime-types');
-const { initializePodProvider } = require('./initialize');
+const { connectPodProvider, clearAllData } = require('./initialize');
 
 jest.setTimeout(80000);
 
@@ -9,19 +9,19 @@ const NUM_PODS = 1;
 
 describe('Test pods creation', () => {
   let actors = [],
-    broker,
+    podProvider,
     alice,
     projectUri;
 
   beforeAll(async () => {
-    broker = await initializePodProvider();
+    await clearAllData();
+
+    podProvider = await connectPodProvider();
 
     for (let i = 1; i <= NUM_PODS; i++) {
-      broker[i] = broker;
-
       const actorData = require(`./data/actor${i}.json`);
-      const { webId } = await broker[i].call('auth.signup', actorData);
-      actors[i] = await broker[i].call(
+      const { webId } = await podProvider.call('auth.signup', actorData);
+      actors[i] = await podProvider.call(
         'activitypub.actor.awaitCreateComplete',
         {
           actorUri: webId,
@@ -30,7 +30,7 @@ describe('Test pods creation', () => {
         { meta: { dataset: actorData.username } }
       );
       actors[i].call = (actionName, params, options = {}) =>
-        broker[i].call(actionName, params, {
+        podProvider.call(actionName, params, {
           ...options,
           meta: { ...options.meta, webId, dataset: actors[i].preferredUsername }
         });
@@ -40,7 +40,7 @@ describe('Test pods creation', () => {
   }, 80000);
 
   afterAll(async () => {
-    await broker.stop();
+    await podProvider.stop();
   });
 
   test('Alice WebID has the required informations', async () => {
