--- tests/services/contacts.app.js
@@ -1,3 +1,0 @@
-const { ContactsApp } = require('@activitypods/contacts');
-
-module.exports = ContactsApp;
