--- tests/services/profiles.app.js
@@ -1,3 +1,0 @@
-const { ProfilesApp } = require('@activitypods/profiles');
-
-module.exports = ProfilesApp;
